Move ProfilePage and navPaths back
